重写安装软件逻辑

+++ Menus/Install.sh
@@ -9,15 +9,22 @@ fi
 mkdir -p "$path"
 rm -rf "$path/linuxtoo"
 
-wget https://g.lsy22.com/lsy/LinuxTool/archive/master.tar.gz -O "${path}/aa" &> /dev/null || echo "脚本下载失败" && exit
-tar -zxf "$path/aa" -C $path &> /dev/null || echo "脚本解压" && exit
+wget https://g.lsy22.com/lsy/LinuxTool/archive/master.tar.gz -O "${path}/aa" &> /dev/null
+
+if [[ -d "${path}/aa" ]];then
+  echo "脚本下载失败"
+  exit
+fi
+tar -zxf "$path/aa" -C $path &> /dev/null
 rm "$path/aa"
-chmod +x "$path/linuxtool/Run.sh" &> /dev/null || echo "脚本添加执行权限失败" && exit
+chmod +x "$path/linuxtool/Run.sh" &> /dev/null
 
 sed -i '/alias tool.*/d' "${HOME}/.bashrc"
 echo "alias tool='$path/linuxtool/Run.sh'" | cat >> "${HOME}/.bashrc"
 
 source "${HOME}/.bashrc"
+export tool="$path/linuxtool/Run.sh"
+
 echo "工具箱已经安装成功"
 echo "位置：${path}/linuxtool"
 echo "命令：tool"
